Rename importJsonBlocks -> importJson.
