+++ src/actions/importText.ts
@@ -18,7 +18,7 @@ import appendToPath from '../util/appendToPath'
 import createId from '../util/createId'
 import head from '../util/head'
 import htmlToJson from '../util/htmlToJson'
-import importJson from '../util/importJsonBlocks'
+import importJson from '../util/importJson'
 import isMarkdown from '../util/isMarkdown'
 import isRoot from '../util/isRoot'
 import markdownToText from '../util/markdownToText'
