+++ src/util/__tests__/roamJsonToBlocks.ts
@@ -7,7 +7,7 @@ import { HOME_PATH, HOME_TOKEN } from '../../constants'
 import exportContext from '../../selectors/exportContext'
 import hashThought from '../../util/hashThought'
 import removeHome from '../../util/removeHome'
-import importJson from '../importJsonBlocks'
+import importJson from '../importJson'
 import initialState from '../initialState'
 import keyValueBy from '../keyValueBy'
 import roamJsonToBlocks, { RoamPage } from '../roamJsonToBlocks'
